заменил библиотеку mobx, добавил ts компоненты.
