+++ package.json
@@ -41,7 +41,7 @@
     "jest-watch-typeahead": "0.4.2",
     "mini-css-extract-plugin": "0.9.0",
     "mobx": "^5.15.6",
-    "mobx-react": "^6.3.0",
+    "mobx-react-lite": "^2.2.2",
     "nes.css": "^2.3.0",
     "optimize-css-assets-webpack-plugin": "5.0.3",
     "pnp-webpack-plugin": "1.6.4",
