--- src/App.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import './App.css';
-import { Playground, Panel } from './components/table';
-
-function App() {
-  return (
-    <div className="App">
-      <Playground />
-      <Panel />
-    </div>
-  );
-}
-
-export default App;
