--- src/components/forest/index.js
@@ -1,2 +1,0 @@
-export { Chunk } from './chunk';
-export { Rabbit } from './rabbit';
