--- src/components/forest/rabbit.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export const Rabbit = () => {
-    return(
-        <div className='rabbit' tabIndex='1' />
-    )
-}
