+++ src/components/panel/index.ts
@@ -1,0 +1,3 @@
+export * from './info';
+export * from './button';
+export * from './panel';
