+++ src/components/playgroud/rabbit.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const Rabbit:React.FC = ():JSX.Element => {
+    return(
+        <span className='rabbit' />
+    )
+}
