--- src/components/table/index.js
@@ -1,4 +1,0 @@
-export { Info } from './info';
-export { Button } from './button';
-export { Panel } from './panel';
-export { Playground } from './playground';
