+++ src/index.js
@@ -1,18 +1,18 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Provider } from 'mobx-react';
 
 import './index.css';
 import App from './App';
 
-
 import stores from './stores/index';
+
+export const storesCtx = React.createContext(stores);
 
 ReactDOM.render(
   <React.StrictMode>
-    <Provider { ...stores }>
+    <storesCtx.Provider value={ stores }>
       <App />
-    </Provider>  
+    </storesCtx.Provider>  
   </React.StrictMode>,
   document.getElementById('root')
 );
